맵 에디터 비트표시기능 추가

+++ Assets/Scripts/GamePlay/Note/SavedMeterChangerNoteData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class SavedMeterChangerNoteData : SavedNoteData
 {
-    public float beatPerBar;
+    public int beatPerBar;
     public float beatLengthRate;
 
     public override Note Summon(NoteSummoner summoner, SavedNoteData data)
